requirements.txt
anyconfig==0.10.1
appdirs==1.4.4
architect==0.6.0
arrow==1.2.2
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
binaryornot==0.4.4
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
datapoint==0.9.8
decorator==5.1.1
Deprecated==1.2.13
Django==3.2.12
django-extensions==3.1.5
django-patternfly==0.0.1
django-redis==5.2.0
django-rest-swagger==2.2.0
django-revproxy==0.10.0
django-timescaledb==0.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
dynaconf==3.1.7
executing==0.8.3
fsspec==2022.1.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
inflection==0.5.1
ipython==8.1.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jinja2-time==0.2.0
jmespath==0.10.0
kafka-python==2.0.2
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
mysqlclient==2.1.0
openapi-codec==1.3.2
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
pip-tools==6.5.1
platformdirs==2.5.1
pluggy==0.13.1
poyo==0.5.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
PyJWT==2.6.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-json-logger==2.0.2
python-slugify==6.1.0
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
redis==4.2.2
requests==2.27.1
rope==0.21.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
save==0.1
simplejson==3.18.1
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
toposort==1.7
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
wrapt==1.14.0
